zebrand-backend/pyproject.toml
[tool.poetry]
name = "zebrand-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.12"
SQLAlchemy = "^2.0.32"
pydantic-settings = "^2.4.0"
psycopg2-binary = "^2.9.9"
argon2-cffi = "^23.1.0"
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
casbin = "^1.41.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
httpx = "^0.28.1"
fastapi-cli = "^0.0.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
